feat: 쿠키에서 아이디 추출하는 함수 구현
