+++ kinderpia_client/src/pages/MyPage.tsx
@@ -6,6 +6,7 @@ import MeetingHistory from '../components/MyPage/MeetingHistory';
 import Tab from '../components/MyPage/Tab';
 import '../styles/mypage/MyPage.scss';
 import axios from 'axios';
+import { extractUserIdFromCookie } from '../utils/extractUserIdFromCookie';
 
 const MyPage: React.FC = () => {
   const tabs = ['내 정보', '모임 내역', '나의 리뷰'];
@@ -13,21 +14,14 @@ const MyPage: React.FC = () => {
   const [userId, setUserId] = useState<string | null>(null);
   const [userInfo, setUserInfo] = useState<any>(null); // 유저 정보를 저장할 상태
 
-  // 토큰값 추출하는 임시 코드--> 토큰을 서버에 넘겨주고 통신해야되는게 맞는듯
-  const extractUserIdFromToken = (token: string): string | null => {
-    if (!token) return null;
-
-    try {
-      // 토큰에서 Bearer 부분 제거
-      const tokenWithoutBearer = token.split(' ')[1];
-      const payload = tokenWithoutBearer.split('.')[1]; // 페이로드 부분 가져오기
-      const decodedPayload = JSON.parse(atob(payload)); // Base64 디코딩 후 JSON으로 변환
-      return decodedPayload.sub; // 사용자 ID로 사용할 필드
-    } catch (error) {
-      console.error('토큰에서 사용자 ID 추출 중 오류 발생:', error);
-      return null;
+  // 컴포넌트가 처음 마운트될 때 실행
+  useEffect(() => {
+    const userIdFromToken = extractUserIdFromCookie();
+    if (userIdFromToken) {
+      setUserId(userIdFromToken);
+      fetchUserInfo(userIdFromToken); // 사용자 정보 요청
     }
-  };
+  }, []);
 
   // 사용자 조회 api
   const fetchUserInfo = async (userId: string) => {
@@ -43,18 +37,6 @@ const MyPage: React.FC = () => {
       console.error('유저 정보를 불러오는 중 오류 발생:', error);
     }
   };
-
-  // 컴포넌트가 처음 마운트될 때 실행
-  useEffect(() => {
-    const token = sessionStorage.getItem('token');
-    if (token) {
-      const userIdFromToken = extractUserIdFromToken(token);
-      if (userIdFromToken) {
-        setUserId(userIdFromToken);
-        fetchUserInfo(userIdFromToken); // 사용자 정보 요청
-      }
-    }
-  }, []);
 
   const renderContent = () => {
     switch (activeTab) {
